TMY prediction cleanup
Created some logic to decide if the supplied set of TMY data generates a last FY or last year analysis. This logic also decides which data ranges will be used for actual data, and which will be used to project data to fill in the rest of the timeframe.

Important code changes to MnV (now ver 1.6)
-in ols_model __init__ : changed self.YTD to self.postEnd (more descriptive)
- moved fraction of savings calculation to its own functions and removed it from the ols_model __init__
- modified the tmy gathering and caluclating code so that it is all done more cleanly by repackaging the functions. It is now called in the generate_savings_summary() function since that should be the only time it needs to be called.
 These functions include -
 _generate_savings_intervals
generate_savings_summary()

updated the archive creator to reflect additions
updated template notebook to reflect new functions

- Flipped the  expected - actual calculation to be actual - expected (even though it goes against my instinct). This will result in NEGATIVE numbers being BAD. So -$6,300  is an extra expenditure of 6,300 bucks and that is bad.