PI_Client submodule addition (V1.6.1)
Added Josue's PI_Client v2_1 as submodule

Update mnv.py to create an instance of the pi_client at the global scrope and to use it in all functions inside of mnv.
Updated the mnv notebook to also use the mnv.pi_client instance to grab data for the notebook. This way all pi_clients are the same, and no repeated instances are generated.

Updated the remodel notebook to call for the mnv script and not the mnv14 script